good messges added more secure etc see yourself

+++ src/main/java/com/proj/java/onlineexaminationsystem/controller/QuestionController.java
@@ -61,8 +61,10 @@ public class QuestionController {
         HttpSession session = request.getSession();
         int quiz_id = questionService.getQuestion(id).getQuiz_id().getQuiz_id();
         if (!session.isNew() && session.getAttribute("role").equals("teacher")) {
-//			return "redirect:/";
-            questionService.deleteQuestion(id);
+            if(questionService.getQuestion(id).getQuiz_id().getTeacher_id().getId() == (Integer) session.getAttribute("id")){
+                questionService.deleteQuestion(id);
+            }
+
         }
         return "redirect:/quiz/" + quiz_id;
     }
